优化了代码逻辑。
