+++ MCSL2_Libs/MCSL2_Aria2Controller.py
@@ -1,1 +1,5 @@
-# import aria2p
+from aria2p.client import Client
+
+class Aria2Controller():
+    pass
+# MCSL2_Aria2Client = Client(host="http://localhost", port=60800, secret="")
