+++ MCSL2_Libs/MCSL2_Utils.py
@@ -1,11 +1,11 @@
-from json import load, loads, dump
+from json import load, dump
 from os import mkdir, listdir
 from os import path as ospath
-from typing import Callable, Any, Dict
+from typing import Callable, Any
 
 from PyQt5.QtCore import QThread, pyqtSignal
 from PyQt5.QtWidgets import QDialog
-from requests import get
+
 
 from MCSL2_Libs.MCSL2_AskDialog import Ui_MCSL2_AskDialog
 from MCSL2_Libs.MCSL2_Dialog import Ui_MCSL2_Dialog
@@ -55,74 +55,13 @@ def InitMCSL():
         pass
 
 
-def ParseDownloaderAPIUrl(DownloadSource, DownloadType):
-    UrlPrefix = "https://raw.iqiq.io/LxHTT/MCSLDownloaderAPI/master/"
-    SourceSuffix = ["SharePoint", "Gitee", "luoxisCloud", "GHProxy", "GitHub"]
-    TypeSuffix = [
-        "/JavaDownloadInfo.json",
-        "/SpigotDownloadInfo.json",
-        "/PaperDownloadInfo.json",
-        "/BungeeCordDownloadInfo.json",
-        "/OfficialCoreDownloadInfo.json",
-    ]
-    DownloadAPIUrl = UrlPrefix + SourceSuffix[DownloadSource] + TypeSuffix[DownloadType]
-    DecodeDownloadJsonsSS = DecodeDownloadJsons(DownloadAPIUrl)
-    SubWidgetNames = DecodeDownloadJsonsSS[0]
-    DownloadUrls = DecodeDownloadJsonsSS[1]
-    FileNames = DecodeDownloadJsonsSS[2]
-    FileFormats = DecodeDownloadJsonsSS[3]
-    return SubWidgetNames, DownloadUrls, FileNames, FileFormats
 
 
-def ParseDownloaderAPIUrl1(DownloadSource):
-    UrlPrefix = "https://raw.iqiq.io/LxHTT/MCSLDownloaderAPI/master/"
-    SourceSuffix = ["SharePoint", "Gitee", "luoxisCloud", "GHProxy", "GitHub"]
-    TypeSuffix = [
-        "/JavaDownloadInfo.json",
-        "/SpigotDownloadInfo.json",
-        "/PaperDownloadInfo.json",
-        "/BungeeCordDownloadInfo.json",
-        "/OfficialCoreDownloadInfo.json",
-    ]
-    rv = {}
-    for i in range(len(TypeSuffix)):
-        DownloadAPIUrl = UrlPrefix + SourceSuffix[DownloadSource] + TypeSuffix[i]
-        SubWidgetNames, DownloadUrls, FileNames, FileFormats = DecodeDownloadJsons(DownloadAPIUrl)
-        rv.update({
-            i: dict(zip(("SubWidgetNames", "DownloadUrls", "FileNames", "FileFormats"),
-                        (SubWidgetNames, DownloadUrls, FileNames, FileFormats)))
-        })
-    return {DownloadSource: rv}
 
 
-def DecodeDownloadJsons(RefreshUrl):
-    SubWidgetNames = []
-    DownloadUrls = []
-    FileFormats = []
-    FileNames = []
-    try:
-        DownloadJson = get(RefreshUrl).text
-    except:
-        # Tip = "无法连接MCSLAPI，\n\n请检查网络或系统代理设置"
-        # CallMCSL2Dialog(Tip, isNeededTwoButtons=0)
-        return -2, -2, -2, -2
-    try:
-        PyDownloadList = loads(DownloadJson)["MCSLDownloadList"]
-        for i in PyDownloadList:
-            SubWidgetName = i["name"]
-            SubWidgetNames.insert(0, SubWidgetName)
-            DownloadUrl = i["url"]
-            DownloadUrls.insert(0, DownloadUrl)
-            FileFormat = i["format"]
-            FileFormats.insert(0, FileFormat)
-            FileName = i["filename"]
-            FileNames.insert(0, FileName)
-        return SubWidgetNames, DownloadUrls, FileNames, FileFormats
-    except:
-        # print(DownloadJson)
-        # Tip = "可能解析api内容失败\n\n请检查网络或自己的节点设置"
-        # CallMCSL2Dialog(Tip, isNeededTwoButtons=0)
-        return -1, -1, -1, -1
+
+
+
 
 
 def ReadJsonFile(FileName):
@@ -230,51 +169,3 @@ class JsonSaveThreadFactory:
         return JsonSaveThread(FileName, Data, FinishSlot)
 
 
-class FetchDownloadURLThread(QThread):
-    """
-    用于获取网页内容的线程
-    结束时发射fetchSignal信号，参数为url和data组成的元组
-    """
-    fetchSignal = pyqtSignal(dict)
-
-    def __init__(self, DownloadSource, FinishSlot: Callable = ...):
-        super().__init__()
-        self._id = None
-        self.DownloadSource = DownloadSource
-        self.Data = None
-        if FinishSlot is not ...:
-            self.fetchSignal.connect(FinishSlot)
-
-    def getURL(self):
-        return self.url
-
-    def run(self):
-        self.fetchSignal.emit(ParseDownloaderAPIUrl1(self.DownloadSource))
-
-    def getData(self):
-        return self.Data
-
-
-@Singleton
-class FetchDownloadURLThreadFactory:
-    singletonThread: Dict[int, FetchDownloadURLThread] = {}
-
-    @classmethod
-    def create(cls,
-               downloadSrc: int,
-               _singleton=False,
-               finishSlot=...) -> FetchDownloadURLThread:
-
-        # print({k: v.isRunning() for k, v in cls.singletonThread.items()})
-        if _singleton:
-
-            if downloadSrc in cls.singletonThread and cls.singletonThread[downloadSrc].isRunning():
-                # print("线程已存在，返回已存在的线程")
-
-                return cls.singletonThread[downloadSrc]
-            else:
-                thread = FetchDownloadURLThread(downloadSrc, finishSlot)
-                cls.singletonThread[downloadSrc] = thread
-                return thread
-        else:
-            return FetchDownloadURLThread(downloadSrc, finishSlot)
