+++ Tools/Aria2/aria2.conf
@@ -14,8 +14,6 @@ enable-rpc=true
 rpc-allow-origin-all=true
 rpc-listen-all=true
 event-poll=select
-rpc-listen-port=6800
+rpc-listen-port=60800
 
 force-save=false
-dht-file-path=dht.dat
-dht-file-path6=dht6.dat
